Separate DataStreamFileProcedures to a separate class
