+++ osu.Framework/Audio/Track/TrackBass.cs
@@ -290,83 +290,6 @@ namespace osu.Framework.Audio.Track
 
         public override bool HasCompleted => base.HasCompleted || IsLoaded && !IsRunning && CurrentTime >= Length;
 
-        private class DataStreamFileProcedures
-        {
-            private byte[] readBuffer = new byte[32768];
-
-            private readonly AsyncBufferStream dataStream;
-
-            public FileProcedures BassProcedures => new FileProcedures
-            {
-                Close = ac_Close,
-                Length = ac_Length,
-                Read = ac_Read,
-                Seek = ac_Seek
-            };
-
-            public DataStreamFileProcedures(AsyncBufferStream data)
-            {
-                dataStream = data;
-            }
-
-            private void ac_Close(IntPtr user)
-            {
-                //manually handle closing of stream
-            }
-
-            private long ac_Length(IntPtr user)
-            {
-                if (dataStream == null) return 0;
-
-                try
-                {
-                    return dataStream.Length;
-                }
-                catch
-                {
-                }
-
-                return 0;
-            }
-
-            private int ac_Read(IntPtr buffer, int length, IntPtr user)
-            {
-                if (dataStream == null) return 0;
-
-                try
-                {
-                    if (length > readBuffer.Length)
-                        readBuffer = new byte[length];
-
-                    if (!dataStream.CanRead)
-                        return 0;
-
-                    int readBytes = dataStream.Read(readBuffer, 0, length);
-                    Marshal.Copy(readBuffer, 0, buffer, readBytes);
-                    return readBytes;
-                }
-                catch
-                {
-                }
-
-                return 0;
-            }
-
-            private bool ac_Seek(long offset, IntPtr user)
-            {
-                if (dataStream == null) return false;
-
-                try
-                {
-                    return dataStream.Seek(offset, SeekOrigin.Begin) == offset;
-                }
-                catch
-                {
-                }
-                return false;
-            }
-        }
-
         public double PitchAdjust
         {
             get { return Frequency.Value; }
